rpc/jsonrpc, docs: widen txpool_status counts and describe base fee waiting
The two protobuf counts are uint32, so pending + baseFee wrapped before the
conversion to hexutil.Uint. Widen both operands and describe base fee sub-pool
transactions as nonce-ready and otherwise valid while they wait for the fee
condition to change, rather than executable.

+++ rpc/jsonrpc/txpool_api.go
@@ -87,8 +87,8 @@ func (api *TxPoolAPIImpl) Content(ctx context.Context) (map[string]map[string]ma
 		}
 		addr := gointerfaces.ConvertH160toAddress(reply.Txs[i].Sender)
 		switch reply.Txs[i].TxnType {
-		// Base fee sub-pool transactions are executable and only wait for a lower base fee,
-		// which is what Geth keeps in its pending list, so they belong to the pending bucket.
+		// Base fee sub-pool transactions are nonce-ready and otherwise valid, waiting only for the
+		// base fee to drop, which is what Geth keeps in its pending list, so they are pending here too.
 		case txpoolproto.AllReply_PENDING, txpoolproto.AllReply_BASE_FEE:
 			if _, ok := pending[addr]; !ok {
 				pending[addr] = make([]types.Transaction, 0, 4)
@@ -187,7 +187,7 @@ func (api *TxPoolAPIImpl) Status(ctx context.Context) (map[string]hexutil.Uint, 
 		return nil, err
 	}
 	return map[string]hexutil.Uint{
-		"pending": hexutil.Uint(reply.PendingCount + reply.BaseFeeCount),
+		"pending": hexutil.Uint(uint64(reply.PendingCount) + uint64(reply.BaseFeeCount)),
 		"queued":  hexutil.Uint(reply.QueuedCount),
 	}, nil
 }
